+++ src/lang/ffi/js/time.rs
@@ -6,7 +6,7 @@
 #[cfg(feature = "time")]
 use crate::TimeDelta;
 #[cfg(all(feature = "unsafe_ffi", not(windows)))]
-use crate::js_doc;
+use crate::{js_doc, WebWindow};
 use crate::{Display, impl_trait};
 #[allow(unused_imports)]
 use crate::{Web, js_number, js_uint32};
@@ -92,28 +92,28 @@ impl JsTimeout {
     #[doc = js_doc!("Window", "setTimeout")]
     /// Calls a function after a delay in milliseconds.
     pub fn timeout(callback: extern "C" fn(), delay_ms: js_uint32) -> Self {
-        Web::window_set_timeout(callback, delay_ms) }
+        WebWindow::set_timeout(callback, delay_ms) }
     #[doc = js_doc!("Window", "setInterval")]
     /// Calls a function repeatedly at a fixed interval in milliseconds.
     pub fn interval(callback: extern "C" fn(), interval_ms: js_uint32) -> Self {
-        Web::window_set_timeout(callback, interval_ms) }
+        WebWindow::set_timeout(callback, interval_ms) }
 
     /// Executes JavaScript code immediately.
     /// ## Security Warning
     /// - Avoid passing untrusted input, as this executes arbitrary JS.
     /// - Ensure all evaluated code is **safe and controlled**.
-    pub fn eval(js_code: &str) { Web::window_eval(js_code) }
+    pub fn eval(js_code: &str) { WebWindow::eval(js_code) }
     #[doc = js_doc!("Window", "setTimeout")]
     /// Executes JavaScript code after a delay in milliseconds.
     pub fn eval_timeout(js_code: &str, delay_ms: js_uint32) -> Self {
-        Web::window_eval_timeout(js_code, delay_ms) }
+        WebWindow::eval_timeout(js_code, delay_ms) }
     #[doc = js_doc!("Window", "setInterval")]
     /// Executes JavaScript code repeatedly at a fixed interval in milliseconds.
     pub fn eval_interval(js_code: &str, interval_ms: js_uint32) -> Self {
-        Web::window_eval_interval(js_code, interval_ms) }
+        WebWindow::eval_interval(js_code, interval_ms) }
 
     #[doc = js_doc!("Window", "clearTimeout")]
     #[doc = js_doc!("Window", "clearInterval")]
     /// Cancels a timeout or interval.
-    pub fn clear(self) { Web::window_clear_timeout(self); }
+    pub fn clear(self) { WebWindow::clear_timeout(self); }
 }